+++ CHANGELOG.md
@@ -33,6 +33,7 @@ This update accompanies the release of the [Telegram Bot API 6.6](https://core.t
 ### Changed
  * Removed unused `__destruct()` method from `\TgBotLib\Bot`
  * Added proper construction for `PollOption[]` objects in `TgBotLib\Objects\Telegram > Poll`
+ * Renamed `thumb` field in `TgBotLib\Objects\Telegram > Animation` to `thumbnail`
 
 ## [6.5.2] - 2023-02-26